feat: initial draft for /build page
